changed: split ECL(Integration|Regression)Test classes to separate files
to unify code with the summary comparison classes
